The btpacket.rb file has been renamed using the snake_case convention
