+++ lib/bt_packet.rb
